refactor(memory): consolidate sync source adapters
