+++ src/openhuman/memory_sync/mod.rs
@@ -28,6 +28,5 @@
 
 pub mod composio;
 pub mod mcp;
-pub mod sources;
 pub mod sync_status;
 pub mod workspace;
