--- src/openhuman/memory_sync/sources/mod.rs
@@ -1,7 +1,0 @@
-//! Memory-source sync pipelines.
-//!
-//! Product-owned wrappers around tinycortex source sync facilities.
-
-pub mod audit;
-pub mod github;
-pub mod rebuild;
